Fix homepage production 500 by dropping sharp from the `/` module graph
`fetchRecentlyPlayed` imported `getImageGradient`, which top-level-imported
sharp. After the sharp 0.35 catalog bump, Vercel traced the linux addon
without libvips and every `/` request died on dlopen.

Derive album-art gradients in the browser via canvas instead, and delete the
server-side sharp path. Homepage NFT no longer includes sharp or libvips.
