+++ apps/web/src/app/home/SpotifyCard.tsx
@@ -1,53 +1,14 @@
 import 'server-only';
 
-import type { Track } from '@dg/content-models/spotify/Track';
 import { ContentCard } from '@dg/ui/dependent/ContentCard';
 import { FaIcon } from '@dg/ui/icons/FaIcon';
 import type { SxObject } from '@dg/ui/theme';
 import { faSpotify } from '@fortawesome/free-brands-svg-icons/faSpotify';
 import { Box, Card, Skeleton, Stack } from '@mui/material';
-import type { ReactNode } from 'react';
 import { Suspense } from 'react';
 import { getLatestSong } from '../../services/spotify';
 import { ALBUM_ART_BORDER_RADIUS, ALBUM_ART_DIMENSIONS } from '../spotify/albumArtStyles';
-import { SpotifyCardScrollTracker } from '../spotify/SpotifyCardScrollTracker';
-import { TrackListing } from '../spotify/TrackListing';
-
-type SpotifyCardShellProps = {
-  children: ReactNode;
-  gradient?: string;
-};
-
-const shellContainerSx: SxObject = {
-  isolation: 'isolate',
-  overflow: 'visible',
-  position: 'relative',
-};
-
-const getGradientGlowSx = (gradient: string): SxObject => ({
-  backgroundImage: gradient,
-  borderRadius: 6,
-  filter: 'blur(16px)',
-  inset: -2,
-  opacity: 0.5,
-  pointerEvents: 'none',
-  position: 'absolute',
-  zIndex: 0,
-});
-
-const cardSx: SxObject = {
-  display: 'flex',
-  minWidth: { md: 'auto', xs: 'min(max-content, inherit)' },
-  overflow: 'visible',
-  padding: 2.5,
-  position: 'relative',
-  zIndex: 1,
-};
-
-const getCardSx = (gradient?: string): SxObject => ({
-  ...(gradient ? { backgroundImage: gradient } : {}),
-  ...cardSx,
-});
+import { SpotifyCardWithGradient } from '../spotify/SpotifyCardWithGradient';
 
 const loadingLayoutSx: SxObject = {
   flex: 1,
@@ -83,21 +44,19 @@ const loadingTitleSx: SxObject = {
   marginBottom: 1,
 };
 
-function SpotifyCardShell({ children, gradient }: SpotifyCardShellProps) {
-  return (
-    <Box sx={shellContainerSx}>
-      {gradient ? <Box aria-hidden="true" sx={getGradientGlowSx(gradient)} /> : null}
-      <ContentCard sx={getCardSx(gradient)}>{children}</ContentCard>
-    </Box>
-  );
-}
+const loadingShellSx: SxObject = {
+  display: 'flex',
+  minWidth: { md: 'auto', xs: 'min(max-content, inherit)' },
+  overflow: 'visible',
+  padding: 2.5,
+};
 
 /**
  * Loading skeleton shown during Suspense.
  */
 function SpotifyCardLoading() {
   return (
-    <SpotifyCardShell>
+    <ContentCard sx={loadingShellSx}>
       <Stack sx={loadingLayoutSx}>
         <Stack direction="row" sx={loadingHeaderSx}>
           <Box sx={loadingLogoSx}>
@@ -113,35 +72,20 @@ function SpotifyCardLoading() {
           <Skeleton height={20} width="60%" />
         </Stack>
       </Stack>
-    </SpotifyCardShell>
-  );
-}
-
-/**
- * Shows a card with the latest data from Spotify, wrapped in the
- * gradient shell and scroll tracker.
- */
-function SpotifyCardContent({ track }: { track: Track | null }) {
-  if (!track) {
-    return null;
-  }
-
-  return (
-    <SpotifyCardScrollTracker>
-      <SpotifyCardShell gradient={track.albumGradient}>
-        <TrackListing track={track} />
-      </SpotifyCardShell>
-    </SpotifyCardScrollTracker>
+    </ContentCard>
   );
 }
 
 /**
  * Async data-fetching wrapper. Fetches the latest song server-side
- * and renders the card content.
+ * and renders the client card that derives its gradient from album art.
  */
 async function SpotifyCardAsync() {
   const track = await getLatestSong();
-  return <SpotifyCardContent track={track} />;
+  if (!track) {
+    return null;
+  }
+  return <SpotifyCardWithGradient track={track} />;
 }
 
 /**
