+++ apps/web/src/app/spotify/rgbToHsl.ts
@@ -1,22 +1,17 @@
-import 'server-only';
+export type Rgb = {
+  r: number;
+  g: number;
+  b: number;
+};
 
-export interface Rgb {
-  r: number; // 0-255
-  g: number; // 0-255
-  b: number; // 0-255
-}
-
-export interface Hsl {
-  h: number; // 0-360
-  s: number; // 0-1
-  l: number; // 0-1
-}
+export type Hsl = {
+  h: number;
+  s: number;
+  l: number;
+};
 
 /**
- * Converts RGB (0-255 channels) to HSL (h: 0-360, s: 0-1, l: 0-1)
- *
- * Algorithm: Standard RGB to HSL conversion
- * Reference: https://www.rapidtables.com/convert/color/rgb-to-hsl.html
+ * Converts RGB (0-255 channels) to HSL (h: 0-360, s: 0-1, l: 0-1).
  */
 export function rgbToHsl({ r, g, b }: Rgb): Hsl {
   const red = r / 255;
@@ -26,7 +21,6 @@ export function rgbToHsl({ r, g, b }: Rgb): Hsl {
   const min = Math.min(red, green, blue);
   const l = (max + min) / 2;
 
-  // Achromatic (grayscale)
   if (max === min) {
     return { h: 0, l, s: 0 };
   }
